00_Base_Project_Structure/aiogram_bot_pro_template/pyproject.toml
[project]
name = "aiogram-bot-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiogram-dialog>=2.3.1",
    "alembic>=1.15.2",
    "dynaconf>=3.2.10",
    "fluentogram>=1.1.10",
    "nats-py>=2.10.0",
    "ormsgpack>=1.9.1",
    "psycopg>=3.2.7",
    "psycopg-pool>=3.2.6",
    "pydantic>=2.0.0",
    "taskiq>=0.11.17",
    "taskiq-nats>=0.5.1",
    "taskiq-redis>=1.0.8",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
